+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - FilterTable.DefaultBuilder refactored, nulls used instead of defaults.
 - AbstractComponentBuilder, AbstractButtonBuilder and AbstractTextComponentBuilder refactored, nulls used instead of defaults.
 - Layout specific panel builders added.
+- DefaultInputDialogBuilder.show(Predicate) bug fixed, no longer closes dialog on ok, regardless of the close predicate.
 ### is.codion.framework.db.local
 - MethodTracer.NO_OP_TRACE added, exit() no longer returns null.
 ### is.codion.framework.server